refactor(hippy-react): refactor event code
