+++ framework/js/packages/hippy-react/src/dom/document-node.ts
@@ -30,7 +30,6 @@ class DocumentNode extends ViewNode {
 
   constructor() {
     super();
-
     this.documentElement = new Element('document');
   }
