+++ framework/js/packages/hippy-react/src/event/emitter.ts
