+++ framework/js/packages/hippy-react/src/event/index.ts
@@ -18,26 +18,9 @@
  * limitations under the License.
  */
 
-import HippyEventListener from './listener';
-
-interface EventEmitterRevoker {
-  callback: number | undefined;
-  bindListener?: HippyEventListener | undefined;
-}
-
-class EventEmitterRevoker implements EventEmitterRevoker {
-  constructor(id: number, listener: HippyEventListener) {
-    this.callback = id;
-    this.bindListener = listener;
-  }
-
-  public remove() {
-    if (typeof this.callback !== 'number' || !this.bindListener) {
-      return;
-    }
-    this.bindListener.removeCallback(this.callback);
-    this.bindListener = undefined;
-  }
-}
-
-export default EventEmitterRevoker;
+export { default as EventDispatcher } from './dispatcher';
+export { default as HippyEventHub } from './hub';
+export { default as Event } from './event';
+export { default as HippyEventEmitter } from './emitter';
+export { default as HippyEventListener } from './listener';
+export { default as HippyEventRevoker } from './emitter-revoker';
